+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - ADDED: Support for `obj` external ELF format file sections.
 - ADDED: Support for `obj_lma`, `obj_vma`, `obj_align`
 - ADDED: std::xor extension
+- ADDED: Support for big-endian writes
 - CHANGED: Unify all error codes as ERR_<n>
 - README: Added Brink features overview to the introduction
 - README: Added Brink standard extensions table